+++ EXAMPLE_LEGOS/model_legos/neigh_rvec_compute_local_variables.f.frag
@@ -1,1 +1,1 @@
-    integer i,j,jj,numnei,atom,atom_ret
+    integer(c_int) i,j,jj,numnei,atom,atom_ret
